refactor: ZaaksysteemId niet voor GetZaken, verder overal verplicht https://github.com/Klantinteractie-Servicesysteem/KISS-frontend/pull/842#discussion_r1700335726
