+++ Kiss.Bff.Test/Zaaksysteem/ZaaksysteemProxyTests.cs
@@ -1,13 +1,8 @@
-﻿using System.Net;
-using System.Text.Json.Nodes;
-using Kiss.Bff.Extern.ZaakGerichtWerken.Zaaksysteem;
+﻿using Kiss.Bff.Extern.ZaakGerichtWerken.Zaaksysteem;
 using Kiss.Bff.Extern.ZaakGerichtWerken.Zaaksysteem.Shared;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Primitives;
 using Moq;
-using RichardSzalay.MockHttp;
 
 namespace Kiss.Bff.Test.Zaaksysteem
 {
@@ -15,263 +10,31 @@ namespace Kiss.Bff.Test.Zaaksysteem
     public class ZaaksysteemProxyTests
     {
         [TestMethod]
-        public async Task ZaaksysteemProxy_sets_zaaksysteemId_for_root_level_array_response()
+        public void Proxy_returns_correct_result()
         {
-            var result = await RunHappyFlowTest("http://example.com", "my-path", default, "[{}]");
+            var baseUrl = "http://example.com";
+            var key = "een sleutel van minimaal 16 karakters";
+            var config = new ZaaksysteemConfig(baseUrl, "", key, null, null);
+            var sut = new ZaaksysteemProxy(new[] { config }, Mock.Of<ILogger<ZaaksysteemProxy>>());
 
-            Assert.IsInstanceOfType<OkObjectResult>(result, out var okObjectResult);
-            Assert.IsInstanceOfType<IEnumerable<JsonNode>>(okObjectResult.Value, out var nodes);
-            var list = nodes.ToList();
-            Assert.AreEqual(1, list.Count);
-            var node = list[0];
-            Assert.AreEqual("http://example.com/", node["zaaksysteemId"]?.ToString());
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_sets_zaaksysteemId_for_paginated_response()
-        {
-            var result = await RunHappyFlowTest("http://example.com", "my-path", default, "{\"results\": [{}]}");
-
-            Assert.IsInstanceOfType<OkObjectResult>(result, out var okObjectResult);
-            Assert.IsInstanceOfType<JsonNode>(okObjectResult.Value, out var rootNode);
-            Assert.IsInstanceOfType<JsonArray>(rootNode["results"], out var nodes);
-            var list = nodes.ToList();
-            Assert.AreEqual(1, list.Count);
-            var node = list[0];
-            Assert.AreEqual("http://example.com/", node?["zaaksysteemId"]?.ToString());
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_sets_zaaksysteemId_for_single_response()
-        {
             var id = Guid.NewGuid();
-            var result = await RunHappyFlowTest("http://example.com", id.ToString(), default, "{}");
+            var result = sut.Get("my-path", baseUrl);
 
+            Assert.IsInstanceOfType<ProxyResult>(result, out var proxyResult);
+            var message = proxyResult.RequestFactory();
+            Assert.IsNotNull(message);
+            Assert.AreEqual(HttpMethod.Get, message.Method);
+        }
+
+        [TestMethod]
+        public void Proxy_returns_500_if_config_is_missing()
+        {
+            var sut = new ZaaksysteemProxy(Enumerable.Empty<ZaaksysteemConfig>(), Mock.Of<ILogger<ZaaksysteemProxy>>());
+            var result = sut.Get("my-path", "");
             Assert.IsInstanceOfType<ObjectResult>(result, out var objectResult);
-            Assert.AreEqual(200, objectResult.StatusCode);
-            Assert.IsInstanceOfType<JsonNode>(objectResult.Value, out var node);
-            Assert.AreEqual("http://example.com/", node?["zaaksysteemId"]?.ToString());
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_uses_ordering_parameter_for_sorting_ascending()
-        {
-            var result = await RunHappyFlowTest("http://example.com", "my-path", "my-column", @"{""results"": [
-                {""my-column"": 2}, {""my-column"": 1}, {""my-column"": 3}
-            ]}");
-
-            Assert.IsInstanceOfType<OkObjectResult>(result, out var okObjectResult);
-            Assert.IsInstanceOfType<JsonNode>(okObjectResult.Value, out var rootNode);
-            Assert.IsInstanceOfType<JsonArray>(rootNode["results"], out var nodes);
-            var list = nodes.Select(x => x!["my-column"]!.GetValue<int>()).ToArray();
-            Assert.AreEqual(3, list.Length);
-            Assert.AreEqual(list[0], 1);
-            Assert.AreEqual(list[1], 2);
-            Assert.AreEqual(list[2], 3);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_uses_ordering_parameter_for_sorting_descending()
-        {
-            var result = await RunHappyFlowTest("http://example.com", "my-path", "-my-column", @"{""results"": [
-                {""my-column"": 2}, {""my-column"": 1}, {""my-column"": 3}
-            ]}");
-
-            Assert.IsInstanceOfType<OkObjectResult>(result, out var okObjectResult);
-            Assert.IsInstanceOfType<JsonNode>(okObjectResult.Value, out var rootNode);
-            Assert.IsInstanceOfType<JsonArray>(rootNode["results"], out var nodes);
-            var list = nodes.Select(x => x!["my-column"]!.GetValue<int>()).ToArray();
-            Assert.AreEqual(3, list.Length);
-            Assert.AreEqual(list[0], 3);
-            Assert.AreEqual(list[1], 2);
-            Assert.AreEqual(list[2], 1);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_proxies_400_status_code_for_lists()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var baseUrl = "http://example.com";
-            var path = "my-path";
-
-            handler.Expect($"{baseUrl}/{path}")
-                .Respond(HttpStatusCode.BadRequest);
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(400, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_proxies_400_status_code_for_single_result()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var baseUrl = "http://example.com";
-            var path = "my-path/" + Guid.NewGuid();
-
-            handler.Expect($"{baseUrl}/{path}")
-                .Respond(HttpStatusCode.BadRequest);
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(400, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_proxies_non_json_for_single_result()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var baseUrl = "http://example.com";
-            var path = "my-path/" + Guid.NewGuid();
-
-            handler.Expect($"{baseUrl}/{path}")
-                .Respond("text/html", "<p>Hello world</p>");
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(200, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_proxies_unreachable_host_for_single_result()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var baseUrl = "http://example.com";
-            var path = "my-path/" + Guid.NewGuid();
-
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(502, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_proxies_unreachable_host_for_lists()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var baseUrl = "http://example.com";
-            var path = "my-path";
-
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(502, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_handles_missing_config_for_single_result()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var path = "my-path/" + Guid.NewGuid();
-
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-
-
-            var sut = new ZaaksysteemProxy(Enumerable.Empty<ZaaksysteemConfig>(), clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(500, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_handles_missing_config_for_lists()
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            var path = "my-path";
-
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-
-
-            var sut = new ZaaksysteemProxy(Enumerable.Empty<ZaaksysteemConfig>(), clientFactory, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(500, statusCodeResult.StatusCode);
-        }
-
-        [TestMethod]
-        public async Task ZaaksysteemProxy_correctly_handles_pagination_mismatch()
-        {
-            var config1 = new ZaaksysteemConfig("http://example1.com", "ClientId", "Secret of at least X characters", null, null);
-            var config2 = new ZaaksysteemConfig("http://example2.com", "ClientId", "Secret of at least X characters", null, null);
-            var path = "my-path";
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-
-            // root level array
-            handler.Expect($"{config1.BaseUrl}/{path}")
-                .Respond("application/json", "[{}]");
-
-            // gepagineerd
-            handler.Expect($"{config2.BaseUrl}/{path}")
-                .Respond("application/json", "{\"results\": [{}]}");
-
-            var clientFactoryMock = new Mock<IHttpClientFactory>();
-            clientFactoryMock.Setup(x => x.CreateClient(It.IsAny<string>())).Returns(() => new HttpClient(handler));
-
-            var sut = new ZaaksysteemProxy(new[] { config1, config2 }, clientFactoryMock.Object, logger);
-
-            var result = await sut.Get(path, null, Enumerable.Empty<string>(), default);
-            Assert.IsInstanceOfType<IStatusCodeActionResult>(result, out var statusCodeResult);
-            Assert.AreEqual(502, statusCodeResult.StatusCode);
-        }
-
-        private static async Task<IActionResult> RunHappyFlowTest(string baseUrl, string path, StringValues ordering, string responseBody)
-        {
-            var logger = Mock.Of<ILogger<ZaaksysteemProxy>>();
-            var handler = new MockHttpMessageHandler();
-            handler.Expect($"{baseUrl}/{path}")
-                .Respond("application/json", responseBody);
-            var httpClient = new HttpClient(handler);
-            var clientFactory = Mock.Of<IHttpClientFactory>(x => x.CreateClient(It.IsAny<string>()) == httpClient);
-            var config = new ZaaksysteemConfig(baseUrl, "ClientId", "Secret of at least X characters", null, null);
-
-
-            var sut = new ZaaksysteemProxy(new[] { config }, clientFactory, logger);
-
-            var result = await sut.Get(path, null, ordering, default);
-
-            return result;
+            Assert.AreEqual(500, objectResult.StatusCode);
+            Assert.IsInstanceOfType<ProblemDetails>(objectResult.Value, out var problemDetails);
+            Assert.AreEqual("Configuratieprobleem", problemDetails.Title);
         }
     }
 }
