+++ src/features/zaaksysteem/components/ZaakDocumenten.vue
@@ -41,24 +41,21 @@ import type { ZaakDetails, ZaakDocument } from "./../types";
 import { Heading as UtrechtHeading } from "@utrecht/component-library-vue";
 import { formatDateOnly } from "@/helpers/date";
 import { formatBytes } from "@/helpers/formatBytes";
-import { fetchLoggedIn, throwIfNotOk } from "@/services";
+import { throwIfNotOk } from "@/services";
 import { Button as UtrechtButton } from "@utrecht/component-library-vue";
+import { fetchWithZaaksysteemId } from "../service";
 
 const props = defineProps<{
   zaak: ZaakDetails;
+  zaaksysteemId: string;
 }>();
-
 // bij het implementeren van meerdere zaaksystemen is gekozen om de zaaksysteemid mee te geven in de header
 // zodat de requests en querystrings verder zo min mogelijk afwijken van de api standaard
 // dit betekent dat we het downloaden van documenten op een omslachtige manier moeten doen,
 // omdat je in een gewone link geen header mee kan geven.
 async function download(doc: ZaakDocument) {
   const url = doc.url + "/download?versie=1";
-  const blob = await fetchLoggedIn(url, {
-    headers: {
-      ZaaksysteemId: props.zaak.zaaksysteemId || "",
-    },
-  })
+  const blob = await fetchWithZaaksysteemId(props.zaaksysteemId, url)
     .then(throwIfNotOk)
     .then((r) => r.blob());
   const objectUrl = URL.createObjectURL(blob);
@@ -66,6 +63,7 @@ async function download(doc: ZaakDocument) {
   a.style.display = "none";
   a.href = objectUrl;
   a.download = doc.bestandsnaam || doc.titel;
+  a.target = "_blank";
   document.body.appendChild(a);
   a.click();
   URL.revokeObjectURL(objectUrl);
